+++ client/src/rpc/traits/cfx.rs
@@ -5,11 +5,12 @@
 use super::super::types::{
     Account as RpcAccount, Block, Bytes, CallRequest, EpochNumber,
     EstimateGasAndCollateralResponse, Filter as RpcFilter, Log as RpcLog,
-    Receipt as RpcReceipt, SponsorInfo as RpcSponsorInfo, Transaction,
-    H160 as RpcH160, H256 as RpcH256, U256 as RpcU256, U64 as RpcU64,
+    Receipt as RpcReceipt, SponsorInfo as RpcSponsorInfo,
+    StorageRoot as RpcStorageRoot, Transaction, H160 as RpcH160,
+    H256 as RpcH256, U256 as RpcU256, U64 as RpcU64,
 };
 use crate::rpc::types::BlockHashOrEpochNumber;
-use jsonrpc_core::{BoxFuture, Result as RpcResult};
+use jsonrpc_core::{BoxFuture, Result as JsonRpcResult};
 use jsonrpc_derive::rpc;
 
 /// Cfx rpc interface.
@@ -17,29 +18,29 @@ use jsonrpc_derive::rpc;
 pub trait Cfx {
     //        /// Returns protocol version encoded as a string (quotes are
     // necessary).        #[rpc(name = "cfx_protocolVersion")]
-    //        fn protocol_version(&self) -> RpcResult<String>;
+    //        fn protocol_version(&self) -> JsonRpcResult<String>;
     //
     /// Returns the number of hashes per second that the node is mining with.
     //        #[rpc(name = "cfx_hashrate")]
-    //        fn hashrate(&self) -> RpcResult<RpcU256>;
+    //        fn hashrate(&self) -> JsonRpcResult<RpcU256>;
 
     //        /// Returns block author.
     //        #[rpc(name = "cfx_coinbase")]
-    //        fn author(&self) -> RpcResult<RpcH160>;
+    //        fn author(&self) -> JsonRpcResult<RpcH160>;
 
     //        /// Returns true if client is actively mining new blocks.
     //        #[rpc(name = "cfx_mining")]
-    //        fn is_mining(&self) -> RpcResult<bool>;
+    //        fn is_mining(&self) -> JsonRpcResult<bool>;
 
     /// Returns current gas price.
     #[rpc(name = "cfx_gasPrice")]
-    fn gas_price(&self) -> RpcResult<RpcU256>;
+    fn gas_price(&self) -> JsonRpcResult<RpcU256>;
 
     /// Returns highest epoch number.
     #[rpc(name = "cfx_epochNumber")]
     fn epoch_number(
         &self, epoch_number: Option<EpochNumber>,
-    ) -> RpcResult<RpcU256>;
+    ) -> JsonRpcResult<RpcU256>;
 
     /// Returns balance of the given account.
     #[rpc(name = "cfx_getBalance")]
@@ -51,7 +52,7 @@ pub trait Cfx {
     #[rpc(name = "cfx_getAdmin")]
     fn admin(
         &self, addr: RpcH160, epoch_number: Option<EpochNumber>,
-    ) -> BoxFuture<RpcH160>;
+    ) -> BoxFuture<Option<RpcH160>>;
 
     /// Returns sponsor information of the given contract
     #[rpc(name = "cfx_getSponsorInfo")]
@@ -83,34 +84,39 @@ pub trait Cfx {
         &self, addr: RpcH160, pos: RpcH256, epoch_number: Option<EpochNumber>,
     ) -> BoxFuture<Option<RpcH256>>;
 
+    #[rpc(name = "cfx_getStorageRoot")]
+    fn storage_root(
+        &self, address: RpcH160, epoch_num: Option<EpochNumber>,
+    ) -> JsonRpcResult<Option<RpcStorageRoot>>;
+
     /// Returns block with given hash.
     #[rpc(name = "cfx_getBlockByHash")]
     fn block_by_hash(
         &self, block_hash: RpcH256, include_txs: bool,
-    ) -> RpcResult<Option<Block>>;
+    ) -> JsonRpcResult<Option<Block>>;
 
     /// Returns block with given hash and pivot chain assumption.
     #[rpc(name = "cfx_getBlockByHashWithPivotAssumption")]
     fn block_by_hash_with_pivot_assumption(
         &self, block_hash: RpcH256, pivot_hash: RpcH256, epoch_number: RpcU64,
-    ) -> RpcResult<Block>;
+    ) -> JsonRpcResult<Block>;
 
     /// Returns block with given epoch number.
     #[rpc(name = "cfx_getBlockByEpochNumber")]
     fn block_by_epoch_number(
         &self, epoch_number: EpochNumber, include_txs: bool,
-    ) -> RpcResult<Block>;
+    ) -> JsonRpcResult<Block>;
 
     /// Returns best block hash.
     #[rpc(name = "cfx_getBestBlockHash")]
-    fn best_block_hash(&self) -> RpcResult<RpcH256>;
+    fn best_block_hash(&self) -> JsonRpcResult<RpcH256>;
 
     /// Returns the nonce should be filled in next sending transaction from
     /// given address at given time (epoch number).
     #[rpc(name = "cfx_getNextNonce")]
     fn next_nonce(
         &self, addr: RpcH160, epoch_number: Option<BlockHashOrEpochNumber>,
-    ) -> RpcResult<RpcU256>;
+    ) -> JsonRpcResult<RpcU256>;
 
     //        /// Returns the number of transactions in a block with given hash.
     //        #[rpc(name = "cfx_getBlockTransactionCountByHash")]
@@ -124,17 +130,17 @@ pub trait Cfx {
 
     /// Sends signed transaction, returning its hash.
     #[rpc(name = "cfx_sendRawTransaction")]
-    fn send_raw_transaction(&self, raw_tx: Bytes) -> RpcResult<RpcH256>;
+    fn send_raw_transaction(&self, raw_tx: Bytes) -> JsonRpcResult<RpcH256>;
 
     //        /// @alias of `cfx_sendRawTransaction`.
     //        #[rpc(name = "cfx_submitTransaction")]
-    //        fn submit_transaction(&self, Bytes) -> RpcResult<RpcH256>;
+    //        fn submit_transaction(&self, Bytes) -> JsonRpcResult<RpcH256>;
 
     /// Call contract, returning the output data.
     #[rpc(name = "cfx_call")]
     fn call(
         &self, tx: CallRequest, epoch_number: Option<EpochNumber>,
-    ) -> RpcResult<Bytes>;
+    ) -> JsonRpcResult<Bytes>;
 
     /// Returns logs matching the filter provided.
     #[rpc(name = "cfx_getLogs")]
@@ -150,17 +156,17 @@ pub trait Cfx {
     #[rpc(name = "cfx_estimateGasAndCollateral")]
     fn estimate_gas_and_collateral(
         &self, request: CallRequest, epoch_number: Option<EpochNumber>,
-    ) -> RpcResult<EstimateGasAndCollateralResponse>;
+    ) -> JsonRpcResult<EstimateGasAndCollateralResponse>;
 
     #[rpc(name = "cfx_getBlocksByEpoch")]
     fn blocks_by_epoch(
         &self, epoch_number: EpochNumber,
-    ) -> RpcResult<Vec<RpcH256>>;
+    ) -> JsonRpcResult<Vec<RpcH256>>;
 
     #[rpc(name = "cfx_getSkippedBlocksByEpoch")]
     fn skipped_blocks_by_epoch(
         &self, epoch_number: EpochNumber,
-    ) -> RpcResult<Vec<RpcH256>>;
+    ) -> JsonRpcResult<Vec<RpcH256>>;
 
     #[rpc(name = "cfx_getTransactionReceipt")]
     fn transaction_receipt(
@@ -177,13 +183,13 @@ pub trait Cfx {
     #[rpc(name = "cfx_getInterestRate")]
     fn interest_rate(
         &self, epoch_number: Option<EpochNumber>,
-    ) -> RpcResult<RpcU256>;
+    ) -> JsonRpcResult<RpcU256>;
 
     /// Returns accumulate interest rate of the given epoch
     #[rpc(name = "cfx_getAccumulateInterestRate")]
     fn accumulate_interest_rate(
         &self, epoch_number: Option<EpochNumber>,
-    ) -> RpcResult<RpcU256>;
+    ) -> JsonRpcResult<RpcU256>;
 
     //        /// Returns transaction at given block hash and index.
     //        #[rpc(name = "cfx_getTransactionByBlockHashAndIndex")]