+++ core/src/executive/internal_contract/impls/admin.rs
@@ -14,7 +14,7 @@ use std::str::FromStr;
 
 lazy_static! {
     pub static ref ADMIN_CONTROL_CONTRACT_ADDRESS: Address =
-        Address::from_str("6060de9e1568e69811c4a398f92c3d10949dc891").unwrap();
+        Address::from_str("8060de9e1568e69811c4a398f92c3d10949dc891").unwrap();
 }
 
 pub fn suicide(