+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 0.4.0
+
+## Bug Fixes
+
+- Fix a potential crash bug in the transaction pool compoenent
+
+- Stop marking OverlayAccount dirty on read access. This will influnce the state root. 
+
+## Improvements
+
+- Improve the transaction address check at RPC
+
+# 0.3.2
+
+## Bug Fixes
+
+- Fix an issue that GetBlockHashesByEpoch containing blocks before checkpoint may cause the node to crash.
+
+- Quick fix for possible duplicate block inserted into the consensus worker thread.
+
 # 0.3.0
 
 ## Blockchain Core Updates (Not Backward Compatible)